+++ CHANGELOG.md
@@ -16,8 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `list_dbt_cloud_run_artifacts` task - [#23](https://github.com/PrefectHQ/prefect-dbt/pull/23)
 - `get_dbt_cloud_run_artifact` task - [#23](https://github.com/PrefectHQ/prefect-dbt/pull/23)
 - `call_dbt_cloud_administrative_api_endpoint` task - [#25](https://github.com/PrefectHQ/prefect-dbt/pull/25)
+- `SnowflakeTargetConfigs` block - [#27](https://github.com/PrefectHQ/prefect-dbt/pull/27)
 
 ### Changed
+- `TargetConfigs` and `GlobalConfigs` moved to `cli/configs/base.py` and converted into a `Block` - [#27](https://github.com/PrefectHQ/prefect-dbt/pull/27)
+- Updated tests to be compatible with core Prefect library (v2.0b9) and bumped required version - [#28](https://github.com/PrefectHQ/prefect-dbt/pull/28)
 
 ### Deprecated